+++ app/src/components/Confirm/ConfirmForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './ConfirmForm.vue'
